Updated with changes from OmniGroup svn [r207477]
General:
- Fixed build issues with Xcode 5.1's version of clang
- Fixed 64-bit issues for iOS
- Fixed deprecated API issues for requiring the 10.9 SDK
- Be more explicit about __autoreleasing in various out-parameter places for compatiblity with '-fobjc-arc-disable-writeback'
- Enabled GCC_WARN_UNDECLARED_SELECTOR in some projects; will continue to advance to others

Configurations:
- Require the 10.9 SDK
- Remove explicit list of architectures for iOS, allowing 64-bit builds

OmniBase:
- Add some more ARC helpers
- Remove some GC helpers

OmniFoundation:
- Improve error handling in OFLockFile
- Add keyed access for OFBijection
- Started moving away from OFSubjectTargetingScriptCommand.
- Added a helper for re-evaluating an object specifier with a different container
- Added more scripting debug helpers
- Improved ARC friendliness of a bunch of code
- Added NSString +invalidXMLCharacterSet and +discouragedXMLCharacterSet and deal better with discouraged characters
- Made OFRandom's shared state thread safe instead of just warning about it
- Made OFPreference KVO compliant for objectValue
- Expose support for a reference date in OFRelativeDateParser/OFRelativeDateFormatter

OmniAppKit:
- Do a better job sending feedback with non Mail.app mailers
- Removed the pattern color picker
- Converted OAAboutPanel to xib
- Added an action to open the application scripts folder (for sandboxed app's usage)
- Cleaned up the find API
- Added NSWindow methods to gather blocks and execute them before windows are displayed
- OAContextControlDelegate is a formal protocol now
- Added AXShownMenu and AXShowMenu support
- +tintedImage:... now returns a self-updating tinted image

OmniDataObjects:
- Removed the ODO_SUPPORT_UNDO setting from ODOFeatures.h since we always have NSUndoManager.

OmniFileExchange:
- Added a test case for a spurious conflict case

OmniInspector:
- Converted to ARC
- More work on support for sidebar inspectors
- Inspectors are finally NSViewController-based

OmniQuartz:
- Avoid an exception if we get a nil colorspace
- Support fill-and-stroke on the vector output path

OmniSoftwareUpdate:
- Fixed unrecognized selector exception on significant time changes on iOS

OmniUI/iPad:
- Improved handling of app-owned 'special' URLs
- Restored single-swatch rendering to OUIColorAttributeInspectorWell
- Add our own image picker rather than trying to shoehorn Apple's into our navigation controller
- Added OUISegmentedViewController
- Updated look for OUIInspectorOptionWheel

OmniUI/iPad-Document
- Fix a crasher when duplicating an OmniPresence document w/o a preview while offline
- Improved import/export support

+++ Frameworks/OmniFoundation/Keychain/OFCredentials-iOS.m
@@ -1,4 +1,4 @@
-// Copyright 2010-2013 The Omni Group. All rights reserved.
+// Copyright 2010-2014 Omni Development, Inc. All rights reserved.
 //
 // This software may only be used and reproduced according to the
 // terms in the file OmniSourceLicense.html, which should be
@@ -416,6 +416,6 @@ BOOL OFHasTrustForChallenge(NSURLAuthenticationChallenge *challenge)
     OSStatus err = SecTrustEvaluate(trust, &evaluationResult); // NB: May block for long periods (eg OCSP verification, etc)
     if (err == errSecSuccess && evaluationResult == kSecTrustResultProceed)
         return YES;
-    NSLog(@"err:%ld, evaluationResult:%lu", err, (unsigned long)evaluationResult);
+    NSLog(@"err:%ld, evaluationResult:%lu", (long)err, (unsigned long)evaluationResult);
     return NO;
 }
